requirements.txt
asgiref==3.8.1
black==24.4.2
boto3==1.35.64
botocore==1.35.64
Brotli==1.1.0
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.1
crispy-bootstrap4==2024.1
cryptography==43.0.3
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-countries==7.6.1
django-crispy-forms==2.2
django-db-connection-pool==1.2.5
django-db-geventpool==4.0.6
django-environ==0.11.2
django-extensions==3.2.3
django-sslserver==0.22
django-storages==1.14.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
graphviz==0.20.3
greenlet==3.0.3
gunicorn==22.0.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.6
psycopg2==2.9.9
pycparser==2.22
pydotplus==2.0.2
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.31
sqlparams==6.0.1
sqlparse==0.5.0
stripe==11.3.0
tomli==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
whitenoise==6.7.0
